Updates and documentation. GNL replaced

+++ srcs/lst/ft_lstsize.c
@@ -6,16 +6,16 @@
 /*   By: fclivaz <fclivaz@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/17 19:12:41 by fclivaz           #+#    #+#             */
-/*   Updated: 2024/11/05 00:42:50 by fclivaz          ###   LAUSANNE.ch       */
+/*   Updated: 2025/01/10 20:21:07 by fclivaz          ###   LAUSANNE.ch       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
 	t_list	*temp;
-	int		size;
+	size_t	size;
 
 	if (!lst)
 		return (0);
